requirements.txt
numpy==1.24.3
tqdm
opencv-python
torchvision==0.17.2
torch==2.2.2
catboost
pillow
pandas
natsort 
timm==0.6.12 
pretrainedmodels==0.7.4 
efficientnet_pytorch==0.7.1 
albumentations==1.3.1